SYNKRON-276 Added getAllOrder with lineItems and getOrder with lineItems

+++ src/main/java/io/github/dftrakesh/cloverrest/model/Order/OrderElement.java
@@ -15,6 +15,6 @@ public class OrderElement {
     private Boolean isVat;
     private String state;
     private Long createdTime;
-    private Double modifiedTime;
+    private Long modifiedTime;
     private ListOrderLineItems lineItems;
 }
